refactor: remove router module and integrate routing directly into service stacks
